fix #280 and cleanup identifiers syntax
Remove \' and \" from syntax table modifications and remove
syntax-propertize-via-font-lock as it breaks font-locking
strings.

Remove defadvice around font-lock-fontify-keywords-region
because it breaks all modes that make use of this function.

Add a new function php-unescape-identifiers to c-lang-defconst
c-before-font-lock-functions to modify syntax table entries for
backslashes in identifiers. As a consequence cc-engine is able
to recognize fully qualified identifiers in the form of \A\B\C
and also \A\B::method()